+++ grr_pygame.py
@@ -62,7 +62,7 @@ class Renderer():
       groccolor = groc.World.BLUE
     else:
       groccolor = groc.World.RED
-    distanceFromTarget = theGroc.world.findDistance(theGroc.x, 
+    distanceFromTarget = theGroc.world.findDistanceXY(theGroc.x, 
                                  theGroc.y, theGroc.targetX, 
                                  theGroc.targetY) 
     hunger = theGroc.hungerThreshold - theGroc.fp 
@@ -135,6 +135,7 @@ class Renderer():
 
 #render.tick
   def tick(self):
+    pygame.display.set_caption(str(self.world.population) + " Grocs")
     pygame.display.flip()
     oldColor = self.worldColor
     self.worldColor = self.world.getWorldColor()
@@ -150,11 +151,11 @@ class Renderer():
         x, y = event.pos
         nearestGroc = self.world.findGrocNearXY(x, y) 
         nearestFood = self.world.findFoodNearXY(x, y)
-        gdist = self.world.findDistance(x, y, 
+        gdist = self.world.findDistanceXY(x, y, 
                                         nearestGroc.x, nearestGroc.y)
-        fdist = self.world.findDistance(x, y, 
+        fdist = self.world.findDistanceXY(x, y, 
                                         nearestFood.x, nearestFood.y)
-        if gdist > nearestGroc.personalRadius:
+        if gdist > nearestGroc.getPersonalRadius():
           pass
         else:
           print(nearestGroc.identify())